+++ Source/Nautilus.DomainModel/ValueObjects/Price.cs
@@ -25,7 +25,7 @@ namespace Nautilus.DomainModel.ValueObjects
         /// </summary>
         /// <param name="value">The price value.</param>
         /// <param name="precision">The price decimal precision.</param>
-        private Price(decimal value, int precision)
+        private Price(decimal value, ushort precision)
             : base(value, precision)
         {
             Condition.NotNegativeDecimal(value, nameof(value));
@@ -59,7 +59,7 @@ namespace Nautilus.DomainModel.ValueObjects
         /// <param name="value">The price value.</param>
         /// <param name="precision">The price precision.</param>
         /// <returns>A <see cref="Price"/>.</returns>
-        public static Price Create(decimal value, int precision)
+        public static Price Create(decimal value, ushort precision)
         {
             return new Price(value, precision);
         }
